Created library for portfolio and asset related functions shared by both implementations
