+++ par_portfolio.c
@@ -1,37 +1,7 @@
 #include <stdio.h>
-#include <gsl/gsl_randist.h>
-#include <gsl/gsl_rng.h>
-#include <gsl/gsl_histogram.h>
-#include <time.h>
-#include <math.h>
+#include <stdlib.h>
 #include <omp.h>
-
-/* Stores a pair of correlated random normal variables */
-struct corr_norm {
-    double var1;
-    double var2;
-};
-
-/* Stores information on the annual distribution of returns of a single risky
- * asset along with its weight within the portfolio */
-struct risky_asset {
-    double mean;
-    double sigma;
-    double port_weight;
-};
-
-/* Returns a struct containing two correlated, normally distributed random
- * variables given the correlation coeffcient between them*/
-struct corr_norm two_corr_norm_rvars(double corr);
-
-/* Calculates the return of a portfolio for one month given distribution
- * information for two risky assets and two correlated random variables */
-double one_month_portfolio_return(struct risky_asset asset1,
-        struct risky_asset asset2, double rand_var1, double rand_var2);
-
-/* Calculates the return of a single risky asset given distribution
- * information for that asset and a random variable for that month */
-double one_month_asset_return(double mean, double sigma, double rand_var);
+#include "portfolio_lib.h"
 
 int main(int argc, char **argv) {
 
@@ -110,41 +80,4 @@ int main(int argc, char **argv) {
     }
 
     exit(0);
-}
-
-struct corr_norm two_corr_norm_rvars(double corr) {
-
-    gsl_rng *r = gsl_rng_alloc(gsl_rng_taus2);
-    // TODO: Randomize the seed properly rather than using the clock
-    clock_t seed = clock();
-    gsl_rng_set(r, seed);
-    double corr_var1 = gsl_ran_ugaussian(r);
-
-    /* Implementing formula for finding two correlated random variables */
-    double corr_var2 = corr * corr_var1 +
-        sqrt(1-pow(corr,2)) * gsl_ran_ugaussian(r);
-
-    gsl_rng_free(r);
-
-    struct corr_norm res;
-    res.var1 = corr_var1;
-    res.var2 = corr_var2;
-    
-    return res;
-}
-
-double one_month_portfolio_return(struct risky_asset asset1,
-        struct risky_asset asset2, double rand_var1, double rand_var2) {
-    return asset1.port_weight *
-            one_month_asset_return(asset1.mean, asset1.sigma, rand_var1) +
-            asset2.port_weight *
-            one_month_asset_return(asset2.mean, asset2.sigma, rand_var2);
-}
-
-double one_month_asset_return(double mean, double sigma, double rand_var) {
-    /* 1/12 is the change in time for monthly returns;
-     * static because it only needs to be calculated once for all
-     * runs of this function */
-    static double delta_t = 1.0/12.0;
-    return mean * delta_t + sigma * sqrt(delta_t) * rand_var;
 }
